pyproject.toml
[tool.poetry]
name = "my-scailar-outbound-backend"
version = "0.1.0"
description = ""
authors = ["hfwittmann <hfwittmann@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
azure-communication-callautomation = "1.4.0b1"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
google-cloud-texttospeech = "^2.26.0"
google-cloud-speech = "^2.32.0"
requests = "^2.32.3"
pydantic = "^2.11.4"
pyaudio = "^0.2.14"
google-genai = "^1.13.0"
python-multipart = "^0.0.20"
msal = "^1.32.3"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
